Update the server handling, configuration and HTTP header functionalities
This commit modifies the functionality of HTTP header validation and parsing in the headers file. It enables the server file to accept client connections and process HTTP requests as well as replaces 'std::vector<std::string>' with the shorthand 'vecStr' in the configuration file. It also includes minor file renaming and removal of unnecessary files like 'socketManager.cpp'.