Refactor Client class constructors and operator assignment
The changes prioritized the refactoring of the Client class in the client.cpp and client.hpp files. Copy constructors and assignment operators were properly defined to ensure the right behavior for copy operations and assignments. Unnecessary member variable _parent_socket was removed and various code cleanups were made to improve readability and efficiency.