Add vector of char pointers and its iterator definitions
The commit introduces two new typing definitions for vectors of char pointers and their iterators in the '0-main.hpp' header file. It promises to streamline handling of character pointers within the vector structures across the source code, leading to more maintainable and readable code. Moreover, some comments have been added in the '0-main.cpp' to denote the start and end points of the main function.