Add getCgiPath function to Location class
The commit introduces a new function getCgiPath in the Location class, allowing to retrieve the cgiPath using the extension. In addition, the function is added to the Location class header file. As a result, it allows easier manipulation and usage of the cgiPath within the Location context which is particularly useful for managing CGI related configurations.